# swagger/swagger.yaml
swagger: "2.0"
info:
  description: "This is a RESTFULL API description of url shorten service document generated by swagger from cocotao.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/)."
  version: "1.0.0"
  title: "Swagger Simple URL Shorten"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "97088693@qq.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "3.122.248.123"
basePath: "/"
schemes:
- "http"
tags:
- name: "url shorten"
  description: "shorten for urls and redirect to them"
  externalDocs:
    description: "Find out more"
    url: "https://github.com/cocotao-sap/urlshortener"
paths:
  /short/item:
    post:
      tags:
      - "short"
      summary: "short a url item and store"
      description: ""
      operationId: "shortItem"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "URL object which should to be shorten"
        required: true
        schema:
          $ref: "#/definitions/URLPayload"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/URLShortenPaylaod"
        201:
          description: ""
          schema:
            $ref: "#/definitions/URLShortenPaylaod"
        400:
          description: "Bad Request: invalid Original Url or Short base Url"
        500:
          description: "Internal Server Error: query URL exist failed or shorten URL failed"
      security:
      - petstore_auth:
        - "write:shorturl"
        - "read:shorturl"
  
  /direct/{code}:
    get:
      tags:
      - "direct"
      summary: "redirect to coded short url"
      description: "coded short url need to be translate to original url and be redirected"
      operationId: "findPetsByStatus"
      produces:
      - "application/json"
      parameters:
      - name: "code"
        in: "path"
        description: "code of short url"
        required: true
        type: "string"
        format: "string"
      responses:
        301:
          description: "Jump to original page when original url found"
        302:
          description: "Jump to error page when original url not found"
        500:
          description: "Internal Server Error: query URL exist failed or redirect error"
      security:
      - petstore_auth:
        - "write:shorturl"
        - "read:shorturl" 
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://oauth2.com"
    flow: "implicit"
    scopes:
      write:shorturl: "insert shorten url into your account"
      read:shorturl: "read stored shorten url in your account"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  URLPayload:
    type: "object"
    properties:
      originalUrl:
        type: "string"
        format: "urlpatten"
      shortBaseUrl:
        type: "string"
        format: "urlpatten"
  URLShortenPaylaod:
    type: "object"
    properties:
      originalUrl:
        type: "string"
        format: "urlpatten"
      shortUrl:
        type: "string"
        format: "urlpatten"
      urlCode:
        type: "string"
        format: "^/[0-9a-z@]{5,15}"
      updatedAt:
        type: "string"
        format: "date"
externalDocs:
  description: "Find out more about Swagger"
  url: "https://swagger.io"